+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@ Release Notes
 * Image Templates: Templates for building and publishing VM images.
 * Virtual Machines: Attach existing managed disks.
 * Storage Accounts: CORS rules respect format of original URI (fixes #1003)
+* WebApps: Enable the docker image tag to be specified for a slot.
 
 ## 1.7.17
 * Container Apps: Fix scaler spelling